Add work in progress Card module
This module will contain the logic for creating, rendering,
and updating "cards", which will display an ngram and monitor
a user's keypresses for progress.  Right now I'm still figuring
out the right way to insulate this into a standalone module and keep other
parts of the application (eg. detecting when the next card should
be shown, creating new cards) in other modules, so I'm working towards a
proof of concept that works, which I'll clean up later.  So right now
this module is a dumping ground for duplicate code, random and
miscellaneous utility functions